feat(UDP-152): Refactor create User

+++ libs/data-access/services/DynamoDbIdentityService.ts
@@ -1,4 +1,9 @@
-import { IdentityInput, IdentityRecordEntity } from '../types/Entity';
+import {
+  CreateUserInput,
+  CreateUserResult,
+  IdentityInput,
+  IdentityRecordEntity,
+} from '../types/Entity';
 import { Logger } from '@libs/utils';
 import { Repository } from '../repositories/Repository';
 import createHttpError from 'http-errors';
@@ -22,7 +27,7 @@ export class DynamoDBIdentityService<T extends IdentityRecordEntity> {
     this.logger = logger;
   }
 
-  public async link(input: IdentityInput) {
+  public async linkIdentity(input: IdentityInput) {
     if (input.appId === input.serviceId) {
       throw createHttpError.BadRequest(
         'AppId and serviceId should not be the same',
@@ -33,16 +38,72 @@ export class DynamoDBIdentityService<T extends IdentityRecordEntity> {
     await this.repository.save(entity);
   }
 
-  public async create(input: IdentityInput) {
-    const result = await this.getById(input.appId, false);
-    if (result) return;
+  public async createAppUser(
+    input: CreateUserInput,
+  ): Promise<CreateUserResult> {
+    const pk = `app#${input.appId}`;
 
-    const entity = await this.createFromInput({
-      ...input,
+    this.logger?.debug('checking is user exists', { pk });
+
+    const exists = await this.repository.getByPk(pk);
+
+    if (exists) {
+      this.logger?.debug('User already exists', { pk });
+      return { udpId: '', created: false };
+    }
+
+    const udpId = uuidv4();
+
+    const entity = {
+      pk,
+      sk: udpId,
+      udpId,
       serviceId: input.appId,
+      serviceName: 'app',
+      ...(input.ttl !== undefined ? { ttl: input.ttl } : {}),
+    } as unknown as T;
+
+    this.logger?.debug('Creating new app user', {
+      pk,
+      sk: udpId,
+      appId: input.appId,
     });
-    this.validateEntity(entity);
+
     await this.repository.save(entity);
+
+    this.logger?.debug('User successfully created', {
+      pk,
+      sk: udpId,
+      appId: input.appId,
+    });
+
+    return { udpId, created: true };
+  }
+
+  public async getByServiceId(
+    serviceName: string,
+    serviceId: string,
+    throwNotFound = true,
+  ) {
+    const pk = `${serviceName.toLowerCase()}#${serviceId}`;
+
+    this.logger?.debug('Getting service user', { pk });
+
+    const user = await this.repository.getByPk(pk);
+
+    console.log({ user });
+
+    if (throwNotFound && !user) {
+      throw createHttpError.NotFound(
+        serviceName === 'app'
+          ? `User not found`
+          : `Service not found ${serviceName} ${serviceId}`,
+      );
+    }
+
+    if (user) {
+      return user;
+    }
   }
 
   public async getById(serviceId: string, throwNotFound = true) {
@@ -67,15 +128,19 @@ export class DynamoDBIdentityService<T extends IdentityRecordEntity> {
     return result;
   }
 
-  public async deleteById(serviceId: string) {
+  public async deleteById(serviceName: string, serviceId: string) {
     if (!serviceId) {
       throw createHttpError.BadRequest(`A valid identifier must be provided`);
     }
 
-    const identifier = await this.getById(serviceId);
+    const identifier = await this.getByServiceId(serviceName, serviceId);
+
+    if (!identifier) {
+      throw createHttpError.NotFound();
+    }
 
     const result = await this.repository.delete({
-      pk: PK_CONSTANT,
+      pk: identifier.pk,
       sk: identifier.sk,
     } as Partial<T>);
 
@@ -92,31 +157,17 @@ export class DynamoDBIdentityService<T extends IdentityRecordEntity> {
         'Missing required field serviceId or appId',
       );
     }
-    if (input.appId === input.serviceId) {
-      // Assume this is the first record of type
-      const udpId = uuidv4();
-      return {
-        pk: PK_CONSTANT,
-        sk: `${input.serviceId}/${udpId}`,
-        lsi_1: `${udpId}/${input.serviceId}`,
-        udpId: uuidv4(),
-        ...(input.ttl ? { ttl: input.ttl } : {}),
-      } as T;
-    }
-
     // look up the udp id by app id
-    const appIdentifier = await this.getById(input.appId);
+    const appIdentifier = await this.getByServiceId('app', input.appId);
+
+    console.log({ appIdentifier });
 
     return {
-      pk: PK_CONSTANT,
-      sk: `${input.serviceId}/${appIdentifier.udpId}`,
-      lsi_1: `${appIdentifier.udpId}/${input.serviceId}`,
+      pk: `${input.serviceName}#${input.serviceId}`,
+      sk: appIdentifier.udpId,
       serviceName: input.serviceName,
       serviceId: input.serviceId,
       udpId: appIdentifier.udpId,
-      accessToken: input.accessToken,
-      idToken: input.idToken,
-      refreshToken: input.refreshToken,
       ...(input.ttl ? { ttl: input.ttl } : {}),
     } as T;
   }
